wip for client traversal to avoid redundant server queries
